refactor:페이지 컴포넌트 파일구조 변경
